Add `linkExpires` option for notifications config
